+++ packages/tracing-internal/src/index.ts
@@ -13,7 +13,11 @@ export {
 export type { LazyLoadedIntegration } from './node';
 
 export {
+  // eslint-disable-next-line deprecation/deprecation
   BrowserTracing,
+  browserTracingIntegration,
+  browserTracingStartNavigationSpan,
+  browserTracingStartPageLoadSpan,
   BROWSER_TRACING_INTEGRATION_ID,
   instrumentOutgoingRequests,
   defaultRequestInstrumentationOptions,
